requirements.txt
GitPython
python-dotenv
pydantic-ai